+++ CHANGELOG.rst
@@ -6,3 +6,34 @@ Changelog
 ------------------
 
 * First release on PyPI.
+
+
+0.0.1 (2022-06-06)
+------------------
+
+* Added person detection.
+
+0.0.2 (2022-06-11)
+------------------
+
+* Changed person detection to support video file input.
+* Fixed naming conventions.
+* Removed extraneous elements.
+
+0.0.3 (2022-06-14)
+------------------
+
+* Added support for both live camera input and video input.
+* Added depth for live camera input.
+
+0.0.4 (2022-07-26)
+------------------
+
+* Used Streamlit to add a basic user interface with the ability to modify input source, video file, model file, and confidence threshold.
+* Included a video preview in the UI.
+
+0.0.5 (2022-09-10)
+------------------
+
+* Added NetworkTables functionality.
+* Made final modifications.